+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 - Fix deprecation raised when serializing callable in certain circumstances (#821)
 
+## 2.2.2 (2019-10-10)
+
+- Fix handling of fifth argument in the error handler (#892)
+- Catch exception from vendors in `Sentry\Transport\HttpTransport` (#899)
+
+## 2.2.1 (2019-09-23)
+
+- Disable default deprecation warning `Sentry\Transport\HttpTransport` (#884)
+
 ## 2.2.0 (2019-09-23)
 
 - Change type hint for both parameter and return value of `HubInterface::getCurrentHub` and `HubInterface::setCurrentHub()` methods (#849)